requirements.txt
需求:

    根据成都项目的调研，可能会从政府取得纸质版的财务数据报表，
    目前我们仿照实际情况整理了两种可能需要解析的财务报表样本，
    一种是直接纸质扫描而来的pdf图片，还有一种是我们直接使用手机拍照得到的图片，
    请组织同学针对两种不同的情况进行图片解析的研发编程工作，
    争取正确的将财务数据从图片中提取出来

依赖:
    python3
    pypdf2
    tesseract-ocr(pip install git+https://github.com/jflesch/pyocr.git)
    Pillow
    pyocr
    wand